Snowpack middleware implementation
