--- src/commands/src-file-extension-mapping.ts
@@ -1,14 +1,0 @@
-export default {
-  jsx: 'js',
-  ts: 'js',
-  tsx: 'js',
-  vue: 'js',
-  svelte: 'js',
-  mdx: 'js',
-  php: 'html',
-  md: 'html',
-  ejs: 'html',
-  njk: 'html',
-  scss: 'css',
-  less: 'css',
-};
