+++ src/middleware/send-error.ts
@@ -1,0 +1,6 @@
+import http from 'http';
+
+export default function sendError(res: http.ServerResponse, status: number) {
+  res.writeHead(status);
+  res.end();
+}
