+++ src/middleware/should-proxy.ts
@@ -1,0 +1,7 @@
+import http from 'http';
+import url from 'url';
+
+export default function shouldProxy(pathPrefix: string, req: http.IncomingMessage) {
+  const reqPath = decodeURI(url.parse(req.url!).pathname!);
+  return reqPath.startsWith(pathPrefix);
+}
